Add app name and logo to top header bar in home screen
Adding the ImfrisivMail icon to the top header bar, next to the hamburger menu button, showcases a distinct identity that we need to strengthen to make users proud of being ImfrisivMail users.

Or something. I dunno I'm not a marketing student. But it does look nice though, and is fairly consistent with the web interfaces of Gmail, Proton Mail and Yahoo Mail.

(Also added is the name of the app, next to the said logo, labeled (for now) in black text, in either Cambria, Georgia, or the user's default sans-serif font—in that order. I didn't like the look of Windows and Chromium's default sans-serif, so I added some serif fonts to the options, but the serif default sucks too.